+++ tensor2tensor/rl/trainer_model_based_agent_only.py
@@ -30,6 +30,7 @@ from __future__ import print_function
 from tensor2tensor.bin import t2t_trainer  # pylint: disable=unused-import
 from tensor2tensor.data_generators import gym_problems_specs
 from tensor2tensor.rl import trainer_model_based
+from tensor2tensor.rl import trainer_model_based_params
 
 
 import tensorflow as tf
@@ -50,7 +51,7 @@ def get_simulated_problem_name(game):
 
 
 def main(_):
-  hparams = trainer_model_based.create_loop_hparams()
+  hparams = trainer_model_based_params.create_loop_hparams()
   problem_name = get_simulated_problem_name(hparams.game)
   world_model_dir = FLAGS.world_model_dir
   agent_model_dir = FLAGS.output_dir
